% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{build.null.model}
\alias{build.null.model}
\title{Applies linear of logistic regregression to the data.}
\usage{
build.null.model(y, x, reg.family = "binomial", verbose = FALSE)
}
\arguments{
\item{y}{A vector with values of dependent variable (outcome).}

\item{x}{A data.frame of covariates.}

\item{reg.family}{A regression family. 
Can be either "binomial" or "gaussian."}

\item{verbose}{Indicates verbosing output. Default: FALSE.}
}
\value{
A list of two: "S" - a dataframe with predictors and "fit"
- an object returned by "glm" function.
}
\description{
Applies linear of logistic regregression to the data.
}
