% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqt-methods.R
\name{geneTestOne}
\alias{geneTestOne}
\title{get.reg.family}
\usage{
geneTestOne(
  phenotype,
  genotype,
  covariates,
  STT = 0.2,
  weight = FALSE,
  cumvar.threshold = 75,
  method = "pca",
  out.type = "D",
  scaleData = FALSE,
  penalty = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{phenotype}{phenotype}

\item{genotype}{genotype}

\item{covariates}{covariates}

\item{STT}{STT}

\item{weight}{weight}

\item{cumvar.threshold}{cumvar.threshold}

\item{method}{method}

\item{out.type}{out.type}

\item{scaleData}{scaleData}

\item{penalty}{penalty}

\item{verbose}{verbose}
}
\value{
rslt
}
\description{
get.reg.family
}
