% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqt-methods.R
\name{get.a}
\alias{get.a}
\title{Get a given STT}
\usage{
get.a(L, STT = 0.2)
}
\arguments{
\item{L}{TODO}

\item{STT}{Numeric indicating soft truncation threshold (STT) to convert 
to gamma parameter (must be <= 0.4).}
}
\value{
a TODO
}
\description{
Get a given STT
}
