% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessLASSO}
\alias{preprocessLASSO}
\title{preprocessLASSO}
\usage{
preprocessLASSO(data, pheno, reg.family, penalty = 0.001)
}
\arguments{
\item{data}{data}

\item{pheno}{pheno data}

\item{reg.family}{reg.family}

\item{penalty}{penalty
Default: FALSE.}
}
\value{
list(S, fit, model)
}
\description{
preprocessLASSO
}
