% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessPCA}
\alias{preprocessPCA}
\title{preprocessPCA}
\usage{
preprocessPCA(data, scaleData, cumvar.threshold, verbose)
}
\arguments{
\item{data}{data}

\item{scaleData}{scaled data}

\item{cumvar.threshold}{cumvar.threshold}

\item{verbose}{Indicates verbosing output
Default: FALSE.}
}
\value{
list(S, indices, model).
}
\description{
preprocessPCA
}
