% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessPLS}
\alias{preprocessPLS}
\title{preprocessPLS}
\usage{
preprocessPLS(data, pheno, scaleData, cumvar.threshold, out.type)
}
\arguments{
\item{data}{data}

\item{pheno}{pheno data}

\item{scaleData}{scaleData}

\item{cumvar.threshold}{cumvar.threshold}

\item{out.type}{out.type
Default: FALSE.}
}
\value{
list(S, Y, model)
}
\description{
preprocessPLS
}
