% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ridge_se}
\alias{ridge_se}
\title{Importing required packages and functions}
\usage{
ridge_se(xs, y, yhat, my_mod, verbose = FALSE)
}
\arguments{
\item{xs}{Genotype matrix}

\item{y}{Phenotype}

\item{yhat}{Ridge/LASSO regression object}

\item{my_mod}{Ridge/LASSO regression object}

\item{verbose}{Indicates verbosing output,
Default: FALSE.}
}
\value{
list(vcov, se). 
vcov: variance-covariance matrix;
se: standard deviation
}
\description{
Importing required packages and functions
}
