% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqt-methods.R
\name{rqt-general}
\alias{rqt-general}
\alias{show.rqt}
\title{General functions of \code{rqt} 
such as accessors and printing.}
\description{
Common methods for class rqt. 
This document lists a series of basic methods for the class rqt
}
\details{
Common methods for class rqt
}
