% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{simple.multvar.reg}
\alias{simple.multvar.reg}
\title{Applies linear of logistic regregression to the data.}
\usage{
simple.multvar.reg(null.model, Z, verbose = FALSE)
}
\arguments{
\item{null.model}{A fitted null model}

\item{Z}{A genotype matrix}

\item{verbose}{Indicates verbosing output. Default: FALSE.}
}
\value{
A list of two: "S" - a dataframe with predictors and "fit"
- an object returned by "glm" function.
}
\description{
Applies linear of logistic regregression to the data.
}
