% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{ridge_se}
\alias{ridge_se}
\title{ridge_se Returns variance-covariance matrix and 
standard deviation from ridge and LASSO
regression fit.}
\usage{
ridge_se(xs, y, yhat, my_mod, verbose = FALSE)
}
\arguments{
\item{xs}{Genotype matrix}

\item{y}{Phenotype}

\item{yhat}{Predicted output}

\item{my_mod}{Ridge/LASSO fit}

\item{verbose}{Indicates verbosing output.
Default: FALSE.}
}
\value{
list(vcov, se). 
vcov: variance-covariance matrix;
se: standard deviation
}
\description{
ridge_se Returns variance-covariance matrix and 
standard deviation from ridge and LASSO
regression fit.
}
